src/A/requirements.txt
boto3>=1.26.138,<1.27.0
beautifulsoup4>=4.12.2,<4.13.0
python-dotenv>=1.0.0,<1.1.0
requests>=2.31.0,<2.32.0
